+++ config/prettier.config.cjs
@@ -1,1 +1,1 @@
-module.exports = { semi: true, singleQuote: true, printWidth: 120, trailingComma: 'all' };
+module.exports = { semi: true, singleQuote: true, printWidth: 120, trailingComma: 'all', endOfLine: 'auto' };